+++ CHANGELOG.md
@@ -74,6 +74,10 @@
 - UI: added playlist revive tool to playlist contextual menu (within 'Other tools' submenu).
 - UI: added loaded status icon setting for UI-only playlists on non-lite mode; disabled by default. See [here](https://hydrogenaudio.org/index.php/topic,120979.msg1082334.html#msg1082334).
 - Exporting Playlist: added new Namida 'Export and convert' preset.
+- External integration: new callbacks.
+	* 'Playlist-Manager-SMP: switch enable panel'		-> { window: string[] }
+	* 'Playlist-Manager-SMP: enable panel'				-> { window: string[] }
+	* 'Playlist-Manager-SMP: disable panel'				-> { window: string[] }
 ### Changed
 - Quick-search: foobar2000 shortcuts are now skipped while quick-searching if JS-host supports it (for ex. to bypass play/pause when pressing space bar).
 - Quick-search: tooltip is now disabled while quick-searching.